requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=0.9.0

